[FEATURE] Add GitHub user namespace "typo3"
In addition to the GitHub namespace "typo3-documentation",
where most of the official TYPO3 documentation and tools
of the TYPO3 Documentation Team are located, with this patch
also the repositories of the TYPO3 Core Team in the
namespace "TYPO3" can be accessed and searched.

Therefore, a new <user> parameter has been added to these
commands with the available values "typo3-documentation",
"typo3" and "all":
```
./bashScripts/get-repos.sh [<type>] [<user>] [<token>]
./bashScripts/grep-for-settings.sh <argument> [<user>]
./bashScripts/search-repos.sh <command> [<user>]
./bashScripts/versionbranch-exist.sh <version> [<user>]
./bashScripts/versionbranch-not-exist.sh <version> [<user>]
```
e.g.
```
./bashScripts/get-repos.sh "all" "all"
./bashScripts/grep-for-settings.sh "project" "typo3-documentation"
./bashScripts/search-repos.sh "grep -rnIE '[Ll]icense' --exclude-dir='.git' --include='composer.json' ." "typo3"
./bashScripts/search-repos.sh "grep -rnIE '[Ll]icense' --exclude-dir='.git' --include='Index.rst' ." "typo3-documentation"
./bashScripts/versionbranch-exist.sh "9.5" "all"
./bashScripts/versionbranch-not-exist.sh "9.5" "typo3-documentation"
```

The namespaces with unique identifier, URL and ignored repositories
are defined in the configuration file `/config.yml` as before.

The related local repositories are fetched to subfolder
`<repodir>/<user>/` instead of the `<repodir>/` folder directly.

Also, the GitHub access token can be passed in all
commands that make GitHub API calls.

Finally, the `versionbranch-exist.sh` and
`versionbranch-not-exist.sh` commands are not limited
to local repositories that start with "TYPO3CMS*".

The PHP files have been streamlined.