[TASK] Remote branches are single source of truth
Local branch names may be obsolete if they were renamed
remotely. The previous implementation expected the same
branch name to be found remotely as locally. This could,
for example, prevent the checkout of a main branch that
was recently renamed from "master" to "main", and checked
out locally as "master".