[FEATURE] Run command recursively with exec-repos.sh (#14)
* [TASK] Force checkout of branches

Branch checkout is enforced at all points to emphasize the batch
processing nature of this tool. No local temporary changes should
prevent the run across all repositories and branches. Changes
should instead be committed immediately after the run, and if
not, they should be disregarded.

* [TASK] Print results also for "Stop on first hit"

Ending the run with "exit 0" stopped the whole tool without printing
the results. Replacing it with "break 3" will only terminate the nested
loops and print the results.

* [TASK] Rename bashScripts/ to bash/

* [TASK] Rename Classes/ to src/

* [FEATURE] Run command recursively with exec-repos.sh

This command traverses all local repositories and branches,
like search-repos.sh does, but allows you to pass any
command and allows to pass commands in a script file.

Run the script by:
```
./bash/exec-repos.sh <command>
```
e.g.
```
./bash/exec-repos.sh "grep -rnIE '\`https://typo3\.org' --exclude-dir='.git' ."
```
or
```
./bash/exec-repos.sh "$(pwd)/command/my-command.sh"
```

Do not forget to first update the local repositories
via get-repos.sh before starting the run.

Both scripts, search-repos.sh and exec-repos.sh,
can probably be merged in the long run.

* [TASK] Add command template "Replace and Push"

This command was used recently to replace all occurrences of
:ref:`<manual>:start` and :ref:`<label> <<manual>:start>` across
all release branches of all repositories during the last
TYPO3 documentation standards run.

* [TASK] Replace search-repos.sh with exec-repos.sh

The command exec-repos.sh is an extension of the command
search-repos.sh, so that the latter is no longer necessary.