[FEATURE] Allow processing of not configured user namespaces
Even if a user namespace, e.g. "gilbertsoft", is not configured
in the /config.yml and /config.local.yml, it can still be
processed in these PHP scripts, if using the new parameter
"force":

- get-contributors.php
- get-repo-branches.php
- get-repo-names.php

This is mainly used in ./bash/modify-repos.sh, which wants to
set repositories of any user as remote target "fork".